+++ src/main/kotlin/com/prasetia/erp/model/cme/CmeProjectDetail.kt
@@ -11,17 +11,20 @@ class CmeProjectDetail(
         val year_project: Long,
         val project_type: String,
         val project_id: String,
-        val nilai_po: Long,
+        val nilai_po: Double,
         val no_po: String?,
-        val nilai_invoice: Long,
-        val nilai_budget: Long,
-        val realisasi_budget: Long,
-        val estimate_po: Long,
+        val nilai_invoice: Double,
+        val nilai_budget: Double,
+        val realisasi_budget: Double,
+        val estimate_po: Double,
         val customer:String?,
         val customer_id:Long?,
         val site_type_id: Long?,
         val area: String?
 ){
-    constructor(): this(0,"",0,"","",0,"",
-            0,0,0,0,"", 0, 0, "")
+    constructor(): this(0,"",0,
+            "","",0.0,
+            "",0.0,0.0,
+            0.0,0.0,"",
+            0, 0, "")
 }
