+++ src/main/kotlin/com/prasetia/erp/model/cme/CmeSummaryYear.kt
@@ -10,11 +10,13 @@ class CmeSummaryYear(
         val year_project: Long,
         val jumlah_site: Long,
         val site_cancel: Long,
-        val nilai_po: Long,
-        val nilai_invoice: Long,
-        val nilai_budget: Long,
-        val realisasi_budget: Float,
-        val estimate_po: Long
+        val nilai_po: Double,
+        val nilai_invoice: Double,
+        val nilai_budget: Double,
+        val realisasi_budget: Double,
+        val estimate_po: Double
 ){
-    constructor(): this(0,0,0,0,0,0,0, 0f, 0)
+    constructor(): this(0,0,0,
+            0,0.0,0.0,
+            0.0, 0.0, 0.0)
 }
