+++ src/main/kotlin/com/prasetia/erp/model/cme/CmeSummaryYearProjectTypeCust.kt
@@ -11,14 +11,17 @@ class CmeSummaryYearProjectTypeCust(
         val jumlah_site: Long,
         val project_type: String,
         val site_cancel: Long,
-        val nilai_po: Long,
-        val nilai_invoice: Long,
-        val nilai_budget: Long,
-        val realisasi_budget: Long,
-        val estimate_po: Long,
+        val nilai_po: Double,
+        val nilai_invoice: Double,
+        val nilai_budget: Double,
+        val realisasi_budget: Double,
+        val estimate_po: Double,
         val site_type_id: Long?,
         val customer:String?,
         val customer_id:Long?
 ){
-    constructor(): this(0,0,0,"",0,0,0,0,0,0, 0, "", 0)
+    constructor(): this(0,0,0,
+            "",0,0.0,
+            0.0,0.0,0.0,
+            0.0, 0, "", 0)
 }
